+++ include/LibGP/compute_avg_edge_length.h
@@ -4,7 +4,7 @@
 namespace LibGP
 {
 	template <typename DerivedV, typename DerivedF>
-	LIBGP_INLINE double compute_avg_edge_length(
+	LIBGP_INLINE Float compute_avg_edge_length(
 		const Eigen::MatrixBase<DerivedV>& V,
 		const Eigen::MatrixBase<DerivedF>& F);
 }
